add a script that generates an emoji data file to use at run time
The raw emoji data is pulled from the emoji-datasource npm package,
which seems to be updated fairly rapidly when new emojis are released.

The generated file is saved in the unicode defined sort order. The
order is kept when we construct the user visible widget, ensuring
similar emojis appear together.

To keep file size reasonable, we only extract the properties we need.